+++ src/Features/WcPayPromotion/Init.php
@@ -102,7 +102,14 @@ class Init {
 		if ( 'no' === get_option( 'woocommerce_show_marketplace_suggestions', 'yes' ) ) {
 			return false;
 		}
+		if ( ! apply_filters( 'woocommerce_allow_marketplace_suggestions', true ) ) {
+			return false;
+		}
 
+		$wc_pay_spec = self::get_wc_pay_promotion_spec();
+		if ( ! $wc_pay_spec ) {
+			return false;
+		}
 		return true;
 	}